Move mocks to legacy
