+++ frontend/test/__support__/server-mocks-legacy/dashboard.ts
