+++ frontend/test/__support__/server-mocks-legacy/search.ts
