+++ frontend/test/__support__/server-mocks-legacy/table.ts
